Move code to module (no changes)
