+++ crates/blockifier/src/transaction.rs
@@ -2,6 +2,7 @@ pub mod account_transaction;
 pub mod constants;
 pub mod errors;
 pub mod objects;
+pub mod post_execution;
 #[cfg(any(feature = "testing", test))]
 pub mod test_utils;
 pub mod transaction_execution;
