+++ crates/blockifier/src/transaction/account_transaction.rs
@@ -13,7 +13,7 @@ use crate::execution::contract_class::ContractClass;
 use crate::execution::entry_point::{
     CallEntryPoint, CallType, EntryPointExecutionContext, ExecutionResources,
 };
-use crate::fee::actual_cost::{ActualCost, ActualCostBuilder, PostExecutionReport};
+use crate::fee::actual_cost::{ActualCost, ActualCostBuilder};
 use crate::fee::fee_utils::verify_can_pay_max_fee;
 use crate::fee::gas_usage::estimate_minimal_fee;
 use crate::retdata;
@@ -25,6 +25,7 @@ use crate::transaction::objects::{
     AccountTransactionContext, HasRelatedFeeType, TransactionExecutionInfo,
     TransactionExecutionResult,
 };
+use crate::transaction::post_execution::PostExecutionReport;
 use crate::transaction::transaction_execution::Transaction;
 use crate::transaction::transaction_types::TransactionType;
 use crate::transaction::transaction_utils::update_remaining_gas;
